pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "soma-core"
version = "0.1.0"
description = "SOMA: Somatic Wisdom Architecture — Wisdom over Memory"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "SOMA Project Team"}
]
keywords = ["agent", "memory", "wisdom", "llm", "rag", "cognitive"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "networkx>=3.0",
    "litellm>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.urls]
Homepage = "https://github.com/soma-project/soma-core"
Repository = "https://github.com/soma-project/soma-core.git"
Issues = "https://github.com/soma-project/soma-core/issues"

[tool.hatch.build.targets.wheel]
packages = ["soma"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
